+++ src/commands/monitor/block.ts
@@ -40,12 +40,13 @@ export default class extends ProfileCommand {
                 console.log('\n');
                 console.log(block);
             }, (err) => {
+                listener.close();
                 err = err.message ? JSON.parse(err.message) : err;
                 console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
             });
         }, (err) => {
-            err = err.message ? JSON.parse(err.message) : err;
-            console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
+            listener.close();
+            console.log(chalk.red('Error'), err);
         });
     }
 }