+++ CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.16.1] - 07-Jan-2020
+## [0.16.1] - 11-Jan-2020
 
 ### Added
 
 - Announce transaction information formatted as a table.
 - Converter module.
 - Offline profile creation.
+- ``monitor all`` command.
+- Synchronous transaction announcement with ``--sync`` option.
 
 ### Changed